requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.8
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==1.5.3
tqdm==4.66.1
pydantic==2.5.0
tiktoken==0.5.2
pydantic-settings
huggingface_hub==0.15.1